Before comm tests
git-svn-id: file:///Users/avp/tmp/svn.20121212/LHPC/mdwf/trunk@259 65d2813f-f11f-0410-b17b-8647a28db151

+++ src/faux/do_AB.c
@@ -1,9 +1,9 @@
 #include <mdwf.h>
 
-long long qx(do_AB)(struct Fermion *r_x,
-		    int size, int Ls,
-		    const struct ABTable *atable,
-		    const struct Fermion *s_x)
+unsigned int qx(do_AB)(struct Fermion *r_x,
+		       int size, int Ls,
+		       const struct ABTable *atable,
+		       const struct Fermion *s_x)
 {
   return 1;
 }
