+++ src/faux/do_ApF.c
@@ -1,13 +1,13 @@
 #include <mdwf.h>
 
-long long qx(do_ApF)(struct Fermion *r_x,
-		     int size, int Ls,
-		     const struct ABTable *atable,
-		     const struct neighbor *neighbor,
-		     const struct SUn *U,
-		     const struct Fermion *s_x,
-		     const struct Fermion *s_y,
-		     void *rb[])
+unsigned int qx(do_ApF)(struct Fermion *r_x,
+			int size, int Ls,
+			const struct ABTable *atable,
+			const struct neighbor *neighbor,
+			const struct SUn *U,
+			const struct Fermion *s_x,
+			const struct Fermion *s_y,
+			void *rb[])
 {
   return 1;
 }
