+++ src/faux/dot_fermion.c
@@ -1,9 +1,9 @@
 #include <mdwf.h>
 
-long long qx(dot_fermion)(double *v_r, double *v_i,
-			  int size, int Ls,
-			  const struct Fermion *a,
-			  const struct Fermion *b)
+unsigned int qx(dot_fermion)(double *v_r, double *v_i,
+			     int size, int Ls,
+			     const struct Fermion *a,
+			     const struct Fermion *b)
 {
   return 1;
 }
