+++ src/faux/madd_fermion.c
@@ -1,9 +1,9 @@
 #include <mdwf.h>
-long long qx(madd_fermion)(struct Fermion *r,
-			   int size, int Ls,
-			   const struct Fermion *a,
-			   double s,
-			   const struct Fermion *b)
+unsigned int qx(madd_fermion)(struct Fermion *r,
+			      int size, int Ls,
+			      const struct Fermion *a,
+			      double s,
+			      const struct Fermion *b)
 {
   return 1;
 }
