+++ src/faux/norm2_fermion.c
@@ -1,7 +1,7 @@
 #include <mdwf.h>
-long long qx(norm2_fermion)(double *v_r,
-			    int size, int Ls,
-			    const struct Fermion *a)
+unsigned int qx(norm2_fermion)(double *v_r,
+			       int size, int Ls,
+			       const struct Fermion *a)
 {
   return 1;
 }
