+++ src/faux/proj_Ug0plus.c
@@ -1,9 +1,9 @@
 #include <mdwf.h>
-long long qx(proj_Ug0plus)(struct ProjectedFermion *r,
-			   int size, int Ls,
-			   const struct up_pack *link,
-			   const struct SUn *U,
-			   const struct Fermion *f)
+unsigned int qx(proj_Ug0plus)(struct ProjectedFermion *r,
+			      int size, int Ls,
+			      const struct up_pack *link,
+			      const struct SUn *U,
+			      const struct Fermion *f)
 {
   return 1;
 }
